+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added language folding support [#1412](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1412)
 - Added reference resolution for `Type` column [#1413](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1413)
 - Added reference resolution for `Target` column [#1414](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1414)
+- Added code completion for common UserRights and Permission elements [#1415](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1415)
 
 ## [2025.1.4]